fix: make interface definitions Solidity spec compliant and fix compatibility issue with `seconds` keyword (#640)
Fixes missing `visibility` on `mint` and `burn`, missing `calldata` (on
input) & `memory` (on output) definition and adjusts `seconds` parameter
of `startRewardCall` to be Solidity compliant; `seconds` is a reserved
keyword.

Ideally the interfaces would be directly imported / iterated on in
`specs`, that would prevent some of these issues as the Solidity
compiler would complain.